Task 16

+++ 16-binary_tree_is_perfect.c
@@ -11,8 +11,8 @@
 
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	int left_height;
-	int right_height;
+	size_t left_height;
+	size_t right_height;
 
 	if (tree == NULL)
 		return (1);
